Add files via upload
Added source code for C++ implementation to filter .fq files based on arbitrary strings in their names. This is geared towards removing kraken2-classified ONT synthetic reads ("kraken:taxid|562") from our long-read files, and requires a number of improvements.

The reads file must be unzipped, and it only writes the filtered file to STOUT.
Use c++ on the OVH, and G++ on the OSX machine.