Fixed excessive filtering bug
Filters for such as 'kraken:taxid|2' would previously filter all taxis that contained the query as a substring. This led to valid sequences tagged with, for example, 'kraken:taxid|234441', being inappropriately removed.

To fix this, I added a check to make sure that the substring being recognized must be at the end of the line [(line.find() + query.length() == line.length().]

This check is a hot-patch, and may not work for all applications.