Added functionality without tsv files
Initially, this script required that a .tsv be generated from another source. Now, running it with just a .fa file as input creates an (unsorted) list of all contigs, to which new names can be added manually.

This tsv can then be used as input in the next run of the program.